// src/app/enums/http-key.ts
export enum RequestUrl{
    testDbUrl = "http://localhost:8080",
    dbUrl = "https://login-pr.up.railway.app",
}

export enum StorageKey{
    accessToken = "access-token",
    refreshToken = "refresh-token",
    username = "username",
    role = "role",
    theme = 'theme',
    lang = "language",
    login = "isLogged"
}